backend/requirements.txt

Flask==3.0.3
Flask-Cors==5.0.0
Flask-PyMongo==2.3.0
Jinja2==3.1.4
pymongo==4.10.1
python-dotenv==1.0.1
Werkzeug==3.0.4
